+++ lib/ruby_saml/metadata.rb
@@ -21,7 +21,7 @@ module RubySaml
     # @return [String] XML Metadata of the Service Provider
     #
     def generate(settings, pretty_print=false, valid_until=nil, cache_duration=nil)
-      meta_doc = XMLSecurity::Document.new
+      meta_doc = RubySaml::XML::Document.new
       add_xml_declaration(meta_doc)
       root = add_root_element(meta_doc, settings, valid_until, cache_duration)
       sp_sso = add_sp_sso_element(root, settings)